Change from 32-bit to 64-bit for int and float in HDF colfile

+++ ImageD11/columnfile.py
@@ -502,9 +502,9 @@ try:
         g.attrs['ImageD11_type'] = 'peaks'
         for t in c.titles:
             if t in INTS:
-                ty = np.int32
+                ty = np.int64
             else:
-                ty = np.float32
+                ty = np.float64
             # print "adding",t,ty
             dat = getattr(c, t).astype( ty )
             if t in list(g.keys()):
@@ -534,9 +534,9 @@ try:
         g.attrs['ImageD11_type'] = 'peaks'
         for t in cf.titles:
             if t in INTS:
-                ty = np.int32
+                ty = np.int64
             else:
-                ty = np.float32
+                ty = np.float64
             g.create_dataset( t, data = getattr(cf, t).astype( ty ) )
         h.close()
